+++ src/Skeletest/Prop.hs
@@ -1,0 +1,11 @@
+module Skeletest.Prop (
+  -- * Settings
+  setDiscardLimit,
+  setShrinkLimit,
+  setShrinkRetries,
+  setConfidence,
+  setVerifiedTermination,
+  setTestLimit,
+) where
+
+import Skeletest.Prop.Internal